Expose: Use Render SetPictureTransform to get decent performances (closes #9).
Before each for each Window Pixmap, an scaled Image was created on the
client-side and then push back to the server. This was obviously really slow
as everything could be done server-side. SHM could have been used but there is
no point as Render handles efficiently scaling.

For now, when the Expose plugin is enabled, the whole screen is repainted,
which will be fixed in later commits.