Expose: Selecting a Window to show through a mouse button was completely broken.
Sending a _NET_ACTIVE_WINDOW ClientMessage to the root Window seems to not be
enough (at least with Awesome and confirmed by wmctrl) when the selected
Window is on a different desktop than the current one, so request to change
current desktop through a _NET_CURRENT_DESKTOP ClientMessage if necessary.

Bump xcb-ewmh required version as window was not given in ClientMessage, thus
it was not working with Awesome 3.4.x as it can manage several root Windows
and use window ClientMessage attribute to determine which root Window it
affects.

Also, add missing requirements for Atoms.