Windows mapped outside the screen and then moved inside were not shown.
I'm a happy user of gnome flashback, and I decided to try your compositor
Unagi. Unfortunately, I found a bug that prevents Unagi to be used with
gnome-panel: when unagi is launched first, and gnome-panel after, the panels
are invisible (but they are sensitive to clicks; they are physically there,
but invisible). If unagi is launched after gnome-panel, all works fine.

I tracked the problem and found that gnome-panel maps each panel outside the
screen, and then scroll them inside; unfortunately, unagi doesn't show a
window that was mapped outside when it is moved inside.

Steps to reproduce with xinit when using only Awesome on Debian unstable
(currently Gnome 3.8):

  0) apt-get install awesome-extra gnome-session-fallback notification-daemon
  1) Create ~/.config/gnome-session/sessions/awesome.session:
     [GNOME Session]
     Name=Awesome
     RequiredComponents=gnome-panel;gnome-settings-daemon;
     RequiredProviders=windowmanager;notifications;
     DefaultProvider-windowmanager=awesome
     DefaultProvider-notifications=notification-daemon

  2) Start X.
  3) Start unagi.
  3) Run ``gnome-session --session awesome''.

Signed-off-by: Arnaud Fontaine <arnau@mini-dweeb.org>