Sometimes window decorations didn't get fully repainted with gnome-player.
It seems to be because gnome-mplayer is just a front-end that embeds mplayer
inside, by passing an X-id. By default, it creates a window with a default
size, but when mplayer is launched, about one second after, it resizes the
window. The problem is that it seems that the window region is not updated, so
only part of the decorations are repainted.

Reviewed-by: Arnaud Fontaine <arnau@debian.org>