Do not consider force repaint time in repaint average time calculation.
This fixes a potential lag issue after a force repaints has been done as the
repaint average time was completely off-chart and anyhow force repaint must be
avoided as much as possible.

Followup of previous commit removing force repaint while Expose is
running. There are still a force repaint on Expose enter/exit, but this should
be fine.